requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# RuralLegalAidBot v2 — Python Dependencies
# Python 3.10+ | Ubuntu 22.04
#
# INSTALL ORDER:
#   1. pip install torch torchvision torchaudio \
#          --index-url https://download.pytorch.org/whl/cu118
#   2. pip install -r requirements.txt
#
# OS-level dependencies (install separately):
#   Ubuntu: sudo apt-get install ffmpeg tesseract-ocr tesseract-ocr-tel
# ─────────────────────────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
aiofiles>=23.2.1

# ── Gemini API ────────────────────────────────────────────────────────────────
# Use the new google-genai SDK (preferred) or old google-generativeai
google-genai>=1.0.0          # new SDK
# google-generativeai>=0.8.0  # old SDK (alternative)

# ── LangChain RAG Stack ───────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
chromadb>=0.5.0

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Reranker — FlashRank cross-encoder (lightweight, local) ──────────────────
flashrank>=0.2.9

# ── Transformers (TTS only — translation moved to Gemini) ─────────────────────
transformers>=4.40.0
accelerate>=0.27.2
sentencepiece>=0.1.99

# ── Speech-to-Text ────────────────────────────────────────────────────────────
openai-whisper>=20231117

# ── Document Extraction ───────────────────────────────────────────────────────
PyMuPDF>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.3.0

# ── Audio I/O ─────────────────────────────────────────────────────────────────
soundfile>=0.12.1
numpy>=1.26.4

# ── Config & Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.1
loguru>=0.7.2

# ── Web Search Agent (free, no API key) ──────────────────────────────────────
duckduckgo-search>=6.0.0

# ── NOT in this file (install manually) ──────────────────────────────────────
# torch==2.1.0  (CUDA-specific — install separately with --index-url)
# llama-cpp-python  (optional, only if using local Mistral fallback)
